:white_check_mark: test(1.0): 添加grpc + spring boot的测试
refs #3
