+++ grpc-samples/grpc-samples-facade/src/main/proto/book.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 option java_multiple_files = true;
 
-package com.github.tonydeng.demo.grpc.facade;
+package com.github.tonydeng.demo.rpc.proto;
 
 enum BookType {
     BOOK = 0;
