+++ grpc-samples/grpc-samples-producer/pom.xml
@@ -27,13 +27,26 @@
             <artifactId>junit-platform-runner</artifactId>
         </dependency>
         <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+        </dependency>
+        <dependency>
             <groupId>com.github.tonydeng.demo.grpc</groupId>
             <artifactId>grpc-samples-facade</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
+            <groupId>com.github.tonydeng.demo.rpc</groupId>
+            <artifactId>utils</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>io.grpc</groupId>
             <artifactId>grpc-stub</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter</artifactId>
         </dependency>
     </dependencies>
 </project>
